+++ Tank/Logic/Block.cpp
@@ -2,7 +2,7 @@
 
 Block::Block() : AbstractEntity() { }
 
-Block::Block(const size_t& posX, const size_t& posY, const size_t& sizeX, const size_t& sizeY) : AbstractEntity(posX, posY, sizeX, sizeY) { }
+Block::Block(const short& posX, const short& posY, const short& sizeX, const short& sizeY) : AbstractEntity(posX, posY, sizeX, sizeY) { }
 
 std::ostream& operator<<(std::ostream& o, const Block& block)
 {
