+++ Tank/Logic/Block.h
@@ -12,5 +12,5 @@ class Block : public AbstractEntity
 public:
 
     Block();
-    Block(const size_t& posX, const size_t& posY, const size_t& sizeX, const size_t& sizeY);
+    Block(const short& posX, const short& posY, const short& sizeX, const short& sizeY);
 };
