+++ Tank/Logic/Map.h
@@ -14,20 +14,20 @@ class Map
 public:
 
     Map();
-    Map(const size_t& sizeX, const size_t& sizeY, std::vector<const Block*> blocks = std::vector<const Block*>());
+    Map(const short& sizeX, const short& sizeY, std::vector<const Block*> blocks = std::vector<const Block*>());
     ~Map();
 
     const std::vector<const Block*>& getBlocks() const;
     const std::vector<Player*>& getPlayers() const;
-    const size_t& getSizeX() const;
-    const size_t& getSizeY() const;
+    const short& getSizeX() const;
+    const short& getSizeY() const;
 
     void add(const Block* const block);
     void add(Player* const player);
 
 private:
 
-    size_t sizeX, sizeY;
+    short sizeX, sizeY;
 
     std::vector<const Block*> blocks;
     std::vector<Player*> players;
